First version of the Project Commlink Database
In future updates, I plan to enhance the project by adding more information and visual elements using CSS and JS. This will include expanding the available details about the characters and incorporating visually appealing designs and layouts. The goal is to create a more immersive and visually appealing experience for users as they explore the Star Wars universe on the wiki.